added subtitles and minor updates

+++ database/migrations/2015_08_16_091346_create_training_table.php
@@ -16,7 +16,7 @@ class CreateTrainingTable extends Migration
           $table->increments('id');
 
           $table->string("training_type"); //INTERNAL / EXTERNAL
-          $table->integer("training_total_cost")->nullable();
+          $table->double("training_total_cost")->nullable();
           $table->text("training_cost_components")->nullable();
           $table->string("training_start_date");
           $table->string("training_end_date");
