+++ database/migrations/2015_08_20_213505_create_income_tax_model_table.php
@@ -16,7 +16,7 @@ class CreateIncomeTaxModelTable extends Migration
             $table->increments('id');
 
             $table -> string("step");
-            $table -> integer("amount_limit");
+            $table -> double("amount_limit");
             $table -> double("rate");
 
             $table->timestamps();
